# modules/core-platform/api/v1/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    CheckoutCommitRequest:
      type: object
      required:
      - cartId
      - customerEmail
      - shippingAddress
      - billingAddress
      - paymentMethodId
      properties:
        cartId:
          type: string
          pattern: \S
        customerEmail:
          type: string
          pattern: \S
        shippingAddress:
          type: string
          pattern: \S
        billingAddress:
          type: string
          pattern: \S
        shippingAmount:
          type: number
        taxAmount:
          type: number
        currency:
          type: string
        paymentMethodId:
          type: string
          pattern: \S
        giftCardCodes:
          type: array
          items:
            type: string
        useStoreCredit:
          type: boolean
    CheckoutResponse:
      type: object
      properties:
        orderId:
          $ref: "#/components/schemas/UUID"
        orderNumber:
          type: string
        paymentIntentId:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        paidAt:
          $ref: "#/components/schemas/OffsetDateTime"
...
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
paths:
...
  /api/v1/checkout/commit:
    post:
      summary: Complete checkout and create order
      description: |
        Atomically:
        1. Validates cart and inventory availability
        2. Creates order record
        3. Processes payment via Stripe
        4. Reduces inventory
        5. Clears cart

        **Idempotent:** Uses X-Idempotency-Key header to prevent duplicate orders.
        **Requires authentication:** Must be logged-in customer or use valid API key.
      tags:
      - Storefront
      parameters:
      - description: Idempotency key (UUID v4) for safe retries
        required: true
        name: X-Idempotency-Key
        in: header
        schema:
          type: string
          pattern: \S
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutCommitRequest"
        required: true
      responses:
        "201":
          description: Order successfully created and payment charged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: "Invalid request (validation errors, out of stock, etc.)"
          content:
            application/json: {}
        "402":
          description: Payment required (payment method declined)
          content:
            application/json: {}
        "409":
          description: Duplicate idempotency key (order already created or in-flight
            request)
          content:
            application/json: {}
        "503":
          description: Service unavailable (checkout feature disabled)
          content:
            application/json: {}
...
  /api/v1/checkout/validate-address:
    post:
      summary: Validate shipping address
      description: |
        Validates a shipping address using USPS or Lob API. Returns normalized address
        if valid, or validation errors if invalid.

        **Use Case:** Call this during checkout step 1 to verify address before proceeding.
        **Feature Flag:** Respects `checkout.address-validation.enabled` flag.
        **Authentication:** Optional - supports both guest and authenticated sessions.
      operationId: validateAddress
      tags:
      - Storefront
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: Address validation completed (may contain errors)
          content:
            application/json: {}
        "400":
          description: Invalid request format
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
...
  /checkout:
    get:
      parameters:
      - name: lang
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
      summary: Checkout
      tags:
      - Storefront Resource
...
info:
  title: Village Storefront API
  version: 1.0.0
  description: Multi-tenant SaaS ecommerce platform API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value